third_party/blink/web_tests/platform/generic/fast/frames/sandboxed-iframe-scripting-04-expected.txt

CONSOLE ERROR: Blocked script execution in 'data:text/html,<script>alert('FAIL: Executed script without allow-scripts in data URL');window.parent.postMessage({'fail': true}, '*');</script>' because the document's frame is sandboxed and the 'allow-scripts' permission is not set.
Verify that sandboxed frames without sandbox='allow-scripts' cannot execute script from data: URLs.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS disallowedExecuted is 0
PASS allowedExecuted is 0
PASS successfullyParsed is true

TEST COMPLETE
